+++ src/redturtle/prenotazioni/contentrules/handlers.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-from plone.app.contentrules.handlers import execute_rules
-
-
-def moved(event):
-    """
-    When a booking is moved, execute the rules assigned to his parent.
-    """
-    execute_rules(event)